refactor(client): one shared scroll spy for admin and settings
